build(deps): update `rustix` to v0.36.16/v0.37.7
This commit updates the proxy's dependency on `rustix` in order to
resolve a potential memory exhaustion issue when using the
`rustix::fs::Dir` iterator with the `linux-raw` backend. This issue is
described in GHSA-c827-hfw6-qwvm.

We currently depend on both `rustix` v0.36 and v0.37 as transitive deps,
so this branch updates the v0.36 dep from v0.36.14 to v0.36.16, and the
v0.37 dependency from v0.37.4 to v0.37.7.

Unfortunately, we weren't able to get Dependabot to bump these deps for
us, because it no longer supports the legacy (non-TOML) `rust-toolchain`
file (see #2487 for details). Therefore, we have to do this bump
manually.