meshtls: Extract TLS id verification out of TLS backends (#2507)
To further prepare for the introduction of Spiffe IDs in the proxy,
this change disables TLS SAN verification in both rustls and boring
TLS implementations. The need to do that stems from the fact that
these implementations only look at DNS SANs and cannot handle
certificates that use URI SANs.

To enable adding this functionality, TLS Id verification is not extracted
as a separate step, after the TLS handshake completes. To do that in
boring, we need to verify_hostname. For rustls, we need to pass a
custom ServerCertVerifier that skips name validation. In our case, the
verifier is nearly identical to ustls::client::WebPkiVerifier, except for
the step where DNS SANs are validated based on the ServerName.

This change is non functional, but it sets the stage to have more control
over TLS Id verification.

Signed-off-by: Zahari Dichev <zaharidichev@gmail.com>