test: Cleanup outbound_balancer_waits_for_ready_endpoint
The outbound_balancer_waits_for_ready_endpoint test doesn't reliably
demonstrate the behavior it is trying to test: we setup two identical
endpoints, so our tests can't be sure that the response is coming from
the correct endpoint.

This change updates this test to provide backends that return varied
responses so we can ensure the desired behavior.

This change also improves the test's logging. The discovery test
infrastructure is updated to allow overriding the logical service
address. This aids disambiguating addresses in logs.