ci: Enable coverage on main and all PRs
In preparation for some larger changes to the proxy, this change enables
codecov on main and on PRs that change code. We need to run coverage on
main so that PRs have a meaningful comparison point.

We use the styfle/cancel-workflow-action to prevent multiple runs
from clogging up Actions Runners.

We can adapt this configuration if it's too annoying, but it seems low
risk to add for now.