metrics: Remove unused features
The aging linkerd-metrics crate includes some features that are unused
in our codebase: a histogram Summary; and a 'Scopes' type. Neither of
these are used in our codebase.

This change removes these defunct features so they need not be
considered as we make future improvements/replacements to our metrics
infrastructure.