metrics: Use prometheus-client for proxy_build_info (#2551)
In preparation to add metrics that report float values, and ultimately in
service of replacing Linkerd's custom metrics implementation with an
upstream library, this change adds an integration with the
prometheus-client crate.

This is accomplished by leaving the existing types and traits in place
and adding a FmtMetrics implementation for the prometheus-client
Registry type, minimizing the change needed to support this new library.

New metrics can use the upstream registry and metric types. The
proxy_build_info metric has been ported to use the new API to validate
the new library. Now, the metrics dump ends with:

    # HELP proxy_build_info Proxy build info.
    # TYPE proxy_build_info gauge
    proxy_build_info{date="2023-12-05T19:52:40Z",git_sha="215c32d3f",profile="release",vendor="code@ver-sea",version="0.0.0-dev.215c32d3f"} 1
    # EOF