app: Setup prom registry for use in balancers
This change updates application initialization to instantiate a
prometheus registry that is then passed (as a reference) to the proxy
stack builders to register metrics during initialization. The registry
is then passed to the admin server's metrics export in an Arc.

This approach is preferable to using a RwLock, as it eliminates a
potential deadlock condition.

No functional changes.