tracing: Ensure that INFO-level spans are preserved
There's a foot-gun when instrumenting asynchronous code with:

    .instrument(tracing::debug_span!(...))

When the debug-level span is not enabled, the code will NOT be
instrumented with INFO-level parent spans. To preserve the parent spans,
we must use:

    .instrument(tracing::debug_span!(...).or_current())

This change updates the proxy's uses of debug_span and trace_span to
ensure that the current span info is preserved. This should improve
diagnostic logging.