identity: Update metrics to follow OpenMetrics best practices
In preparation to add SPIRE integration, this change extracts the identity
certificate metrics from the control plane client, moving them into the
linkerd-identity crate, as an implementation of Credentials trait (which is
updated to include an expiry value). The app::identity module is updated to
instrument the credentials store with metrics.

A `identity_cert_refresh_timestamp_seconds` metric is added to report the time
when the identity certificate was last updated.

**Backwards-incompatible changes**: `identity_cert_refresh_count` is renamed to
`dentity_cert_refreshes_total` to follow OpenMetrics best practices, where
counters must end with `_total`.

I've searched GitHub, and the only external references to this metric are in our
documentation (which will have to be updated).

---

    # HELP identity_cert_expiration_timestamp_seconds Time when the this proxy's current mTLS identity certificate will expire (in seconds since the UNIX epoch).
    # TYPE identity_cert_expiration_timestamp_seconds gauge
    # UNIT identity_cert_expiration_timestamp_seconds seconds
    identity_cert_expiration_timestamp_seconds 1704500155.0
    # HELP identity_cert_refresh_timestamp_seconds Time when the this proxy's current mTLS identity certificate were last updated.
    # TYPE identity_cert_refresh_timestamp_seconds gauge
    # UNIT identity_cert_refresh_timestamp_seconds seconds
    identity_cert_refresh_timestamp_seconds 1704413735.7489427
    # HELP identity_cert_refreshes The total number of times this proxy's mTLS identity certificate has been refreshed by the Identity provider..
    # TYPE identity_cert_refreshes counter
    identity_cert_refreshes_total 1