http: Parameterize NewServeHttp
We plan to add defensive timeouts to the HTTP server to limit idle streams and
connections. In preparation for this, we need to parameterize the server
to accept these additonal configurations.

This change updates NewServeHttp to use an ExtractParam to build a Params struct
for each server. In follow-up changes, the timeout configuration will be
instrumented through this Params struct.

There are no functional changes in this commit.