ci: Automate dependabot merges
We currently run many different workflows for each PR, each flagged on a subset
of files. This means that we can't mark any workflows as required for PRs to
merge, which in turn limits our ability to automate the merge process for
dependencies.

To fix this, we consolidate our various integration testing steps into a single
Pull Request workflow (which will be marked as required for PRs to merge).

Similarly, the coverage workflow now runs on on all PRs; and we use the
changed-files action to limit execution to changes that alter Rust code.

All workflows are updated to use GitHub's native concurrency control, which
replaces styfle/cancel-workflow-action.