ci: Update release workflow to be triggered via API
Previously, we have triggered release workflows by pushing a tag to the repo.
While this isn't a huge problem, it's sub-optimal for automation workflows,
where can rely on the `gh` CLI, and it's a sub-optimal user interface compared
to the GitHub Actions UI.

This change updates the release workflow to be triggered by a workflow_dispatch
event with a set of metadata that is used to drive the release process. These
inputs can be specified when triggering the workflow. For example:

    :; gh workflow run release.yml -f version=v2.220.0 -f publish=true