ci: Update release workflow to use a PAT (#2824)
The proxy release workflow now pushes tags. When pushing a tag, we can be in the
position of modifying GitHub workflows, and so we need 'workflows: write'
permissons. These permisson cannot be granted to github-actions[bot].

This change updates the release worfklow to use Actions metadata to
configure GitHub authentication.