ci: Automate weekly proxy releases
This commit adds a release-weekly workflow that runs on Wednesday nights. If
there hasn't been a proxy release earlier in the week and there are new commits
on main, then a new release is triggered.

Furthermore, the release workflow is udpated to trigger the sync-proxy
workflow in the Linkerd2 repo after a successful release.