proxy-http: Box request bodies
The proxy HTTP server emits requests with an UpgradeBody type (as required by
the server module that sets up HTTP CONNECT handling). We also wrap HTTP/2
requests in this body even though we do not currently support CONNECT over
HTTP/2. The inbound and outbound server stacks then Box the request body so that
the UpgradeBody type is hidden.

This commit moves the Box-ing of the request body into the HTTP server to reduce
boilerplate.

This change sets up for additional HTTP server changes. There are no functional
changes.