proxy-http: Simplify TracingExecutor
The proxy::http::trace module was a bit of a misnomer, as it has nothing to do
with "tracing" per-se, and much more to do with being a hyper executor.

This change renames the module and hides it form public export for clarity. It
also simplifies instantiation of the executor.