chore(app,outbound)!: Decouple metrics registry from stack building (#2887)
As we introduced the newer prometheus-client metrics registry, we did so by
allowing stacks to register metrics directly. This is incongruent with the
existing legacy metrics registries, which are known by the Inbound and Outbound
stack builders.

This leads to two problems:

1. We cannot build the admin server until all of the proxy stacks are built.
   This ordering dependency is unnecessary and cumbersome if we want to insert
   additional discovery work into the initialization process.
2. It is cumbersome to add new metrics to stacks, as the registry must be passed
   through the stack building process.

To fix this, this change introduces additional Metrics types so that these
metrics may be registered along with the other stack metrics.

BREAKING CHANGE: Gateway-mode proxies now report Balancer metrics with the
"outbound_" prefix instead of the "gateway_" prefix. This metrics scope was
introduced very recently and is incongruent with our other metrics export. We
have no known readers of these metrics, as they are only relevant to new load
balancer behavior, and only on multicluster gateways. Unifying under the
outbound_ prefix reduces noise in metrics in non-gateway use cases and makes it
easier to query proxy data consistently.

This is the only functional change in this commit.