chore(http): test HTTP/2 flow control exhaustion (#2895)
This change adds tests that exercises the server's behavior when clients exhaust
their HTTP/2 receive windows, i.e. so that the server is unable to send
additional data to a client.

These tests currently only document the existing behavior, but they will be
extended to validate mitigation strategies as they are implemented.