chore(http): split h2 client and server config (#2915)
Hyper supports a variety of configuration options for for the HTTP/2 client and
servers. The proxy's http client and server types do not support modifying these
parameters. We want to support configuring HTTP/2 server and client keep-alives;
and if we are exposing these parameters to the control plane, it also makes
sense to include the others, especially the per-connection stream concurrency
limit and the initial window sizes.

This commit splits the h2 'Settings' type into 'ClientParams' and
'ServerParams', including the appropriate per-connection settings for each.
These parameters are wired into the hyper client and server builders.

These parameters are not yet exposed to the control plane via configuration
interfaces.