Unwrap IPv4-mapped IPv6 addresses
The policy logic doesn't account for IPv4-mapped IPv6 addresses when matching allowed IPs with client IPs, so this change "unwraps" these types of IPs, converting those sockets into `SocketAddr::V4` at the lowest level in `BindTcp`. This also allows us to simplify the logic introduced by 7338f6b2b27c40ef8748551bed3e89905322f151 into `BindWithOrigDst` as it no longer has to account for IPv4-mapped IPv6 addresses (it was detecting them but wasn't unwrapping them).

As an example, this is the error we were getting, when prometheus (at 10.42.0.11) attempted to scrape a pod, even when that IP was authorized:
```
[    69.592605s]  INFO ThreadId(02) daemon:admin{listen.addr=[::]:4191}: linkerd_app_inbound::policy::http: Request denied server.group=policy.linkerd.io server.kind=server server.name=smoke-test-proxy-admin route.group= route.kind=default route.name=default client.tls=Some(Established { client_id: Some(ClientId(Dns(Name("prometheus.linkerd-viz.serviceaccount.identity.linkerd.custom.domain")))), negotiated_protocol: None }) client.ip=::ffff:10.42.0.11
[    69.592664s]  INFO ThreadId(02) daemon:admin{listen.addr=[::]:4191}:rescue{client.addr=[::ffff:10.42.0.11]:55168}: linkerd_app_core::errors::respond: HTTP/2.0 request failed error=unauthorized request on route
```