fix(stream_timeouts): Fix Error type classification (#3092)
The previous implementation of the stream timeout errors did not properly wrap
error type so they could be discovered by the metrics middleware. This change
fixes the error wrapping so that the underlying error type is discoverable as a
source.

This change also fixes a request header timeout calculation. This feature is not
currently exposed to the router stack.