Vendor OpenTelemetry protocol
In preparation for supporting the OpenTelemetry protocol for the proxy's traces, this vendors the relevant protobufs as well as some convenience helpers here.

This works around the fact that the existing `opentelemetry` crates do not support the version of tonic that we currently use (0.10), and this vendoring can be removed once we update.

[#10111](https://github.com/linkerd/linkerd2/issues/10111)

Signed-off-by: Scott Fleener <scott@buoyant.io>